close #5, clients now have option to set port
boost::program_options was used for providing
command line argument parsing for the text-client.
The gtk-client takes these values from the gui anyway.

The ClientStateMachine had to be changed because it didn't
have a port parameter in the constructor yet. The port used is
also broadcast in the `connecting` GameEvent now.